Text changes

+++ Appabetical/BackupUtils.swift
@@ -9,18 +9,20 @@ import Foundation
 
 // Save a homescreen backup manually
 func saveLayout() {
-    do {
-        if fm.fileExists(atPath: savedLayoutUrl.path) {
-            try fm.removeItem(at: savedLayoutUrl)
+    UIApplication.shared.confirmAlert(title: "Confirm Save", body: "This will overwrite your previously saved layout. Would you like to continue?", onOK: {
+        do {
+            if fm.fileExists(atPath: savedLayoutUrl.path) {
+                try fm.removeItem(at: savedLayoutUrl)
+            }
+            try fm.copyItem(at: plistUrl, to: savedLayoutUrl)
+            // Set modification date to now
+            let attributes: [FileAttributeKey : Any] = [.modificationDate: Date()]
+            try fm.setAttributes(attributes, ofItemAtPath: savedLayoutUrl.path)
+            UIApplication.shared.alert(title: "Layout Saved", body: "Layout has been saved successfully.")
+        } catch {
+            UIApplication.shared.alert(body: error.localizedDescription)
         }
-        try fm.copyItem(at: plistUrl, to: savedLayoutUrl)
-        // Set modification date to now
-        let attributes: [FileAttributeKey : Any] = [.modificationDate: Date()]
-        try fm.setAttributes(attributes, ofItemAtPath: savedLayoutUrl.path)
-        UIApplication.shared.alert(title: "Layout Saved", body: "Layout has been saved successfully.")
-    } catch {
-        UIApplication.shared.alert(body: error.localizedDescription)
-    }
+    }, noCancel: false)
 }
 
 // Restore the manual homescreen backup
